requirements.txt
altgraph==0.17.2
ansicon==1.89.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
charset-normalizer==2.0.12
commonmark==0.9.1
ezenv==0.92
future==0.18.2
idna==3.3
jinxed==1.2.0
MouseInfo==0.1.3
numpy==1.23.0
opencv-python==4.5.3.56
pandas==1.4.3
pathlib==1.0.1
pefile==2022.5.30
Pillow==9.1.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
requests==2.28.0
rich==12.4.4
six==1.16.0
urllib3==1.26.9
wincertstore==0.2
